+++ resources/views/admin/block/index.php
@@ -12,11 +12,15 @@ declare(strict_types=1);
 use Hirtz\Cms\Modules\Admin\Data\BlockActiveDataProvider;
 use Hirtz\Cms\Modules\Admin\Widgets\Grids\BlockGridView;
 use Hirtz\Cms\Modules\Admin\Widgets\Navs\BlockHeader;
+use Hirtz\Skeleton\Modules\Admin\Widgets\HintAlert;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Grids\GridContainer;
 
 echo BlockHeader::make()
     ->provider($provider);
+
+echo HintAlert::make()
+    ->text(Yii::t('cms', 'BLOCK_INDEX_HINT'));
 
 echo GridContainer::make()
     ->grid(BlockGridView::make()